project/requirements.txt
# Optimized Renode Peripheral Model Generator Requirements

python-dotenv
pyyaml
pymilvus
milvus
tiktoken
jsonschema
tenacity
rich
ollama
sentence-transformers
google-generativeai
groq
asyncio
jsonschema==4.22.0
python-slugify==8.0.4
tqdm